Every purpose says why it answers as it does
The rail printed one reason for the whole panel — correspondence's — so a
subject who asked us to stop marketing got a blocked Newsletter row with
nothing under it saying why. A refusal a rep cannot explain to the person
in front of them is not usable, and that sentence was already in the
file, applied to one row out of several.

The reason moves onto ConsentRow, so every purpose carries its own. It
draws only where the row shows a verdict: a sentence explaining an answer
the row above replaced with "no address" explains nothing.

The panel comes out into personconsentpanel.tsx to pay for the lines,
which also gives the rail's narrowest question its own file. It is
deliberately narrower than the composer's: the guard answers per PURPOSE
about a person, the composer answers about a MESSAGE, and a panel that
collapsed them would print a permission with no scope on it and be
contradicted by the drawer a moment later.

1575 lines to 1430, waiver ratcheted.

Two tests, one mutation-verified. The second says the correspondence
reason still shows, so the fix cannot become "print the last reason" and
take the row a reply rides out with it.
